# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Blueverse Apis
  version: 1.0.0
servers:
  - url: https://api.dev.blueverse.foxlabs.in/api/v1
components:
  securitySchemes:
    apikeyAuth:
      type: http
      scheme: apikey
...
paths:
...
  /external/washtype:
    post:
      tags:
        - External
      summary: Sync Organisation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                OrganisationGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                Name: GOLD
                InternalName: Test
                Description: Test Description
                PlcTag: Test Plc
                AddBy: Vinay Kumar
                AddDate: '2023-03-21'
                ExpectedWashTime: 21
                OrderNo: 1
                Prefix: test
                IsValid: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /external/machine:
    post:
      tags:
        - External
      summary: Sync Machine
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MindsphereLinker: Test
                Identifier: 1
                AreaGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineBusinessModeGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                Name: Machine 1
                ShortName: MH1
                MachineTag: MHTag
                FoamPrice: 12.43
                ShampooPrice: 12.43
                WaxPrice: 12.43
                ElectricityPerUnitPrice: 12.43
                WaterPerLiterPrice: 12.43
                Latitude: '12.21113'
                Longitude: '12.323223'
                PlcTag: TestTag
                PhLowerLimit: 12.43
                PhUpperLimit: 12.43
                OilAndGreesLowerLimit: 12.43
                OilAndGreesUpperLimit: 12.43
                TssLowerLimit: 12.43
                TssUpperLimit: 12.43
                CodLowerLimit: 12.43
                CodUpperLimit: 12.43
                IsValid: false
                IsDeleted: false
                AddBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                AddDate: '2023-03-30'
                UpdateBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                UpdateDate: '2023-03-20'
                MindsphereDataParameter: Test
                MindsphereRunTimeParameter: Test
                TDSLowerLimit: 13.21
                TDSUpperLimit: 14.22
                IotAssetId: TestId
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/machine-business-mode:
    post:
      tags:
        - External
      summary: Sync Machine BusinessMode
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                Name: Test Mode
                IsValid: true
                IsDeleted: false
                AddBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                AddDate: '2023-03-02'
                UpdateBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                UpdateDate: '2023-03-30'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /external/machine-status:
    post:
      tags:
        - External
      summary: Sync Machine Status
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                Status: true
                AddDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/machine-status-history:
    post:
      tags:
        - External
      summary: Sync Machine Status History
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                Status: true
                AddDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/machine-status-log:
    post:
      tags:
        - External
      summary: Sync Machine Status Log
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                id: 1
                message: Test Message
                message_template: Test Template
                level: 1
                timestamp: '2023-03-12'
                exception: Text
                log_event: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/machine-parameters:
    post:
      tags:
        - External
      summary: Sync Machine Parameters
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                TSSValue: 12.123
                CODValue: 12.323
                OilAndGreaseValue: 13.2332
                IsValid: true
                IsDeleted: false
                AddBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                AddDate: '2023-03-21'
                UpdateBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                UpdateDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/machine-parameters-audit:
    post:
      tags:
        - External
      summary: Sync Machine Parameters Audit
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Id: a3e339b8-cd5b-11ed-afa1-0242ac120002
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                TSSValue: 12.123
                CODValue: 12.323
                OilAndGreaseValue: 13.2332
                IsValid: true
                IsDeleted: false
                AddBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                AddDate: '2023-03-21'
                UpdateBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                UpdateDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/machines-parameter-upper-lower-audit:
    post:
      tags:
        - External
      summary: Sync Machine Parameters Upper Lower Audit
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Id: a3e339b8-cd5b-11ed-afa1-0242ac120002
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                PhLowerLimit: 12.123
                PhUpperLimit: 12.323
                TDSLowerLimit: 13.2332
                TDSUpperLimit: 13.2332
                TssLowerLimit: 13.2332
                TssUpperLimit: 13.2332
                CodLowerLimit: 13.2332
                CodUpperLimit: 13.2332
                OilAndGreesLowerLimit: 13.2332
                OilAndGreesUpperLimit: 13.2332
                UpdateDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/machine-runtime:
    post:
      tags:
        - External
      summary: Sync Machine Runtime
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineRunTime: 1234
                RunTimeDate: '2023-03-21'
                UpdateDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/machine-runtime-history:
    post:
      tags:
        - External
      summary: Sync Machine Runtime History
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineRunTime: 1234
                RunTimeDate: '2023-03-21'
                UpdateDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/escalation-matrix-machine:
    post:
      tags:
        - External
      summary: Sync Escalation Matrix Machine
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                EscalationContactGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                AddDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/escalation-matrix-parameter:
    post:
      tags:
        - External
      summary: Sync Escalation Matrix Parameter
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                EscalationContactGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                AddDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/escalation-matrix:
    post:
      tags:
        - External
      summary: Sync Escalation Matrix
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MachineGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                AlarmGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                EscalationContactGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                SortOrder: 1
                EscalationInterval: 2
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /external/escalation-contacts:
    post:
      tags:
        - External
      summary: Sync Escalation Contacts
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Guid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                Name: Test
                DesignationGuid: a3e339b8-cd5b-11ed-afa1-0242ac120002
                MobileNumber: '8787878765'
                Email: test@gmail.com
                IsValid: true
                IsDeleted: false
                AddBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                AddDate: '2023-03-21'
                UpdateBy: a3e339b8-cd5b-11ed-afa1-0242ac120002
                UpdateDate: '2023-03-21'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...